Moves list ui files into their own package
