+++ app/src/main/java/com/dmgdavid2109/xapochallenge/kotlinrepos/ui/list/KotlinRepoListViewModelInput.kt
@@ -1,4 +1,4 @@
-package com.dmgdavid2109.xapochallenge.kotlinrepos.ui
+package com.dmgdavid2109.xapochallenge.kotlinrepos.ui.list
 
 import com.dmgdavid2109.xapochallenge.common.ui.LceViewModelInputs
